Proper (?) type for color_ref
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@1218 49c17534-1d02-dc11-b283-0030488c597c

+++ co_vn_graphics.c
@@ -30,7 +30,7 @@ struct{
 	float	*qubes;
 	float	*rays;
 	float	*color_vertex;
-	float	*color_ref;
+	uint	*color_ref;
 	float	*color_color;
 	float	*name;
 	float	*dust;
